fix UID decoder

+++ src/rdm/mdb.c
@@ -107,7 +107,7 @@ int rdm_decode_uids(const rdm_mdb_t *mdb, void *data, int num) {
   if (mdb && mdb->pdl && data) {
     const struct __attribute__((__packed__)) {
       uint16_t manufacturer;
-      uint64_t device;
+      uint32_t device;
     } *pd = (void *)mdb->pd;
     rdm_uid_t *param = data;
     for (int i = 0; i < num && decoded * 6 < mdb->pdl; ++i) {
